Make it possible to add keys to the kernel config.
Prior to this change, it was only possible to modify the values
of a key already in use.

This should make it possible to configure stuff like squashfs,
which only has one of ~five keys present in the default kernel
configuration.

Of course - I decided that squashfs wasn't needed yet, so the
only remnants of that thought are this configuration code...

+++ pkgspecs/i686-tools-linux.make.sh
@@ -41,13 +41,18 @@ set_config()
   echo "Setting 'CONFIG_$key' to '$val'"
 
   pattern='^(CONFIG_'"$key"'=).*$'
-  # TODO: Escape $val here!
-  replace='\1'"$val"
-  expr="s@$pattern@$replace@g"
-  echo "sed expr: $expr" # TODO: delme
-  mv .config .config.tmp
-  sed -re "$expr" .config.tmp > .config
-  rm .config.tmp
+  if grep -E "$pattern" .config
+  then
+    # TODO: Escape $val here!
+    replace='\1'"$val"
+    expr="s@$pattern@$replace@g"
+    echo "sed expr: $expr" # TODO: delme
+    mv .config .config.tmp
+    sed -re "$expr" .config.tmp > .config
+    rm .config.tmp
+  else
+    echo "CONFIG_${key}=$val" >> .config
+  fi
   return 0
 }
